[refactoring/daengle-32] auth 관련 커스텀 예외 추가
