+++ daengle-auth/src/main/java/ddog/auth/exception/AuthException.java
@@ -1,4 +1,6 @@
-package ddog.auth.exception.common;
+package ddog.auth.exception;
+
+import ddog.auth.exception.common.CustomRuntimeException;
 
 public class AuthException extends CustomRuntimeException {
     public AuthException(AuthExceptionType type, Object... args) {
