+++ daengle-user-api/src/main/java/ddog/user/application/auth/AuthService.java
@@ -2,8 +2,8 @@ package ddog.user.application.auth;
 
 import ddog.auth.config.jwt.JwtTokenProvider;
 import ddog.auth.dto.TokenAccountInfoDto;
-import ddog.auth.exception.common.AuthException;
-import ddog.auth.exception.common.AuthExceptionType;
+import ddog.auth.exception.AuthException;
+import ddog.auth.exception.AuthExceptionType;
 import ddog.domain.account.Role;
 import ddog.persistence.port.AccountPersist;
 import ddog.user.presentation.auth.dto.LoginResult;
